DBZ-8582 Cleanup tests
