+++ debezium-connector-binlog/src/test/java/io/debezium/connector/binlog/BinlogVectorIT.java
@@ -5,7 +5,6 @@
  */
 package io.debezium.connector.binlog;
 
-import static io.debezium.junit.EqualityCheck.LESS_THAN;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.nio.file.Path;
@@ -23,25 +22,23 @@ import io.debezium.connector.binlog.util.TestHelper;
 import io.debezium.connector.binlog.util.UniqueDatabase;
 import io.debezium.data.vector.FloatVector;
 import io.debezium.jdbc.JdbcConnection;
-import io.debezium.junit.SkipWhenDatabaseVersion;
 
 /**
  * @author Jiri Pechanec
  */
-@SkipWhenDatabaseVersion(check = LESS_THAN, major = 9, minor = 0, reason = "VECTOR datatype not added until MySQL 9.0")
 public abstract class BinlogVectorIT<C extends SourceConnector> extends AbstractBinlogConnectorIT<C> {
 
-    private String databaseName;
+    private static final Path SCHEMA_HISTORY_PATH = Files.createTestingPath("file-schema-history-json.txt")
+            .toAbsolutePath();
+
+    private final String databaseName;
+
+    protected UniqueDatabase DATABASE;
+    protected Configuration config;
 
     public BinlogVectorIT(final String databaseName) {
         this.databaseName = databaseName;
     }
-
-    private static final Path SCHEMA_HISTORY_PATH = Files.createTestingPath("file-schema-history-json.txt")
-            .toAbsolutePath();
-    protected UniqueDatabase DATABASE;
-
-    protected Configuration config;
 
     @Before
     public void beforeEach() {
@@ -143,4 +140,46 @@ public abstract class BinlogVectorIT<C extends SourceConnector> extends Abstract
 
         stopConnector();
     }
+
+    @Test
+    public void shouldConsumeAllEventsFromDatabaseUsingStreaming() throws SQLException, InterruptedException {
+        // Use the DB configuration to define the connector's configuration ...
+        config = DATABASE.defaultConfig()
+                .with(BinlogConnectorConfig.SNAPSHOT_MODE, BinlogConnectorConfig.SnapshotMode.NO_DATA)
+                .build();
+
+        // Start the connector ...
+        start(getConnectorClass(), config);
+
+        // ---------------------------------------------------------------------------------------------------------------
+        // Consume all of the events due to startup and initialization of the database
+        // ---------------------------------------------------------------------------------------------------------------
+        // Testing.Debug.enable();
+        int numTables = 1;
+        int numDdlRecords = numTables * 2 + 3; // for each table (1 drop + 1 create) + for each db (1 create + 1 drop + 1 use)
+        int numSetVariables = 1;
+        var records = consumeRecordsByTopic(numDdlRecords + numSetVariables);
+
+        try (BinlogTestConnection db = getTestDatabaseConnection(DATABASE.getDatabaseName());) {
+            try (JdbcConnection connection = db.connect()) {
+                connection.execute(getShouldConsumeAllEventsFromDatabaseUsingStreamingSql());
+            }
+            records = consumeRecordsByTopic(1);
+        }
+
+        assertThat(records).isNotNull();
+        final var dataRecords = records.recordsForTopic(DATABASE.topicForTable("dbz_8157"));
+        assertThat(dataRecords).hasSize(1);
+        var record = dataRecords.get(0);
+        var after = ((Struct) record.value()).getStruct("after");
+        assertThat(after.schema().field("f_vector_null").schema().name()).isEqualTo(FloatVector.LOGICAL_NAME);
+        assertThat(after.getArray("f_vector_null")).containsExactly(10.1f, 10.2f);
+        assertThat(after.getArray("f_vector_default")).containsExactly(20.1f, 20.2f);
+        assertThat(after.getArray("f_vector_cons")).containsExactly(30.1f, 30.2f);
+
+        stopConnector();
+    }
+
+    protected abstract String getShouldConsumeAllEventsFromDatabaseUsingStreamingSql();
+
 }
